feat: use authentication
