+++ Controllers/HomeController.cs
@@ -8,29 +8,10 @@ namespace SensorWeb.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly ISensorModelRepository sensorModelRepository;
-        private readonly DataFromAPI dataFromAPI;
-
-        public HomeController(ISensorModelRepository sensorModelRepository, DataFromAPI dataFromAPI)
-        {
-            this.sensorModelRepository = sensorModelRepository;
-            this.dataFromAPI = dataFromAPI;
-        }
 
         public IActionResult Index()
         {
-            IEnumerable<SensorModel> sensorsList = sensorModelRepository.GetAll();
-            return View(sensorsList);
-        }
-
-        [HttpPost, ActionName("Index")]
-        public async Task<ActionResult> IndexPOST()
-        {
-            SensorModelJson temp = await dataFromAPI.GetData();
-            Console.WriteLine(dataFromAPI.ConvertSensorModelJson(temp).Name);
-            sensorModelRepository.Add(dataFromAPI.ConvertSensorModelJson(temp));
-            sensorModelRepository.Save();
-            return RedirectToAction("Index");
+            return View();
         }
 
         public IActionResult Privacy()
