+++ Views/Home/Index.cshtml
@@ -1,30 +1,17 @@
-﻿@{
-    ViewData["Title"] = "Home Page";
-}
+﻿@using Microsoft.AspNetCore.Identity
 
-@model IEnumerable<SensorModel>
+@inject SignInManager<IdentityUser> SignInManager
+@inject UserManager<IdentityUser> UserManager
 
-<div class="text-center">
-    <h1 class="display-4">SensorModel</h1>
-</div>
-
-<form method="post" asp-action="Index">
-    <input type="submit" value="Получить данные" />
-</form>
-
-<div>Sensors в БД</div>
-@if (Model != null)
+@if (SignInManager.IsSignedIn(User))
 {
-    @foreach (var Sensor in Model)
-    {
-        <div>Id: @Sensor.Id</div>
-        <div>Name: @Sensor.Name</div>
-        <div>OwnerName: @Sensor.OwnerName</div>
-        <div>GPS: @Sensor.GPS</div>
-        <div>Glonass: @Sensor.Glonass</div>
-        <div>Сharge: @Sensor.Сharge</div>
-        <div>Indication: @Sensor.Indication</div>
-        <div>Date: @Sensor.Date</div>
-        <hr />
-    }
+    <a class="nav-link text-dark" asp-area="" asp-controller="Sensors" asp-action="Index">Go to Sensors</a>
+}
+else
+{
+    <div class="text-center">
+        <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
+            <partial name="_LoginPartial" />
+        </nav>
+    </div>
 }
